+++ src/test/java/com/campool/service/GearTypeServiceTest.java
@@ -60,7 +60,7 @@ class GearTypeServiceTest {
     @DisplayName("존재하지 않는 ID로 삭제 요청 시 예외 발생")
     @Test
     void deleteNonExistentIdThrowsException() {
-        int nonExistentId = 123;
+        long nonExistentId = 123;
         given(gearTypeMapper.findGearTypeById(nonExistentId)).willReturn(null);
 
         Exception exception = assertThrows(NoSuchElementException.class,
